chore(cleanup): Clean up documentation & re-organize the files for d10 Re-roll functions
