+++ module/api/dice/meta-rolld10.mjs
@@ -1,19 +1,13 @@
-//* Meta Roll Functions that roll d10 dice and apply Damage/healing
-//! DSN doesn't show up on subsequent re-rolls of the metaDamageReRoll - fixed
-//! DSN also doesn't show up on the first re-roll from the metaExecute result - fixed
-//todo need to review the structure in lieu of the new VFX & DSN support (Roll Orchestrator)
-//todo need to order the flow and account for who's able (supposed) to view the VFX and the dice rolls
-//todo need to review the whole anchor/re-roll concept and simplify accross all meta-roll functions  - done
-//todo undoing life changes should return a promise so we can continue rather than waiting 3 sec = done
-//todo clean up notes - 
 /**
  * metaRolld10 handles the rolling of d10 dice for a given actor and purpose.
  *
- * This function determines the number of d10 dice to roll based on the provided parameters.
  * It checks for the presence of certain Metapowers that might affect the roll and then performs the roll.
  * If destinyReRoll is set to true, it allows for a re-roll of that roll result, by spending a Destiny Point.
- *
- * Anchored rolls do not trigger DSN when the function is called, but rather when the chat message is created/updated instead.
+ * For Non-Anchored rolls, metaRolld10 will create (or update if reroll is true) the chat message and trigger DSN animation
+ * For Anchored rolls, metaRolld10 will return the roll to anchor and won't trigger DSN or update any chat message.
+ * For Anchored rolls, the calling function needs to take care of triggering DSN and create/update the chat message.
+ * todo need to review the structure in lieu of the new VFX & DSN support (Roll Orchestrator)
+ * todo need to order the flow and account for who's able (supposed) to view the VFX and the dice rolls
  *
  * @param {Object} actor - The actor performing the roll. Expected to be an Actor object.
  * @param {string} what - The reason or purpose for the roll. Expected to be a string. (eg: "Damage")
@@ -27,6 +21,7 @@
  * @param {number} [rerollCounter=0] - The number of rerolls that have been performed. Expected to be a positive number.
  * @param {string} [messageId=null] - The message ID of the chat message for the reroll, if any. Expected to be a string.
  * @returns {Promise<...>}
+ *
  * @example
  * Rolling an actor's Weapon Damage for 3 * d10:
  * metanthropes.dice.metaRolld10(actor, "Damage", true, 3, "Weapon Name");
@@ -173,14 +168,12 @@ export async function metaRolld10(
 	});
 	//* Enrich the message
 	const enrichedMessage = await foundry.applications.ux.TextEditor.enrichHTML(message, { async: true });
-	//* Early Return for Anchored rolls
+	//* Anchored rolls - we don't manage DSN or Chat
 	if (anchor) {
 		const updatedRoll = JSON.stringify(rolld10.toJSON());
 		const renderedRoll = await rolld10.render();
-
 		mL(3, "metaRolld10", "Anchored", reroll ? "Re-Roll" : "No Re-roll");
 		mL(3, "metaRolld10", "Finished for:", actor.name + "'s", what);
-
 		return rolld10.toAnchor({
 			label: what,
 			dataset: {
@@ -201,19 +194,11 @@ export async function metaRolld10(
 			},
 		});
 	}
-
+	//* Non Anchored rolls - we will manage the DSN & Chat
 	if (!anchor) {
-		//* Not anchored, print message to chat
 		if (!reroll) {
-			//!des parakatw note
 			//* Not a reroll, printing a new message
 			mL(3, "metaRolld10", "Not Anchored", "No Re-Roll", "Creating new chat message");
-			//todo! need to find a way to tell dice so nice to only show the animation if dice > 0
-			//todo oxi message edw? //if ( message?.rolls.length && ("dice3d" in game) ) await game.dice3d.waitFor3DAnimationByMessageID(message.id);
-			// if (game.dice3d && dice <= 0) {
-			// 	rolld10.dice[0].results[0].hidden = true;
-			// 	mL(4, "metaRolld10", "not anchor, no reroll", rolld10);
-			// }
 			const updatedRoll = await rolld10.toJSON();
 			const renderedRoll = await rolld10.render();
 			rolld10.toMessage({
@@ -229,11 +214,9 @@ export async function metaRolld10(
 			mL(3, "metaRolld10", "Not Anchored", "Re-Roll", "Updating chat message", messageId);
 			const chatMessage = game.messages.get(messageId);
 			if (!chatMessage) {
-				ui.notifications.warn("Could not find the chat message to update.");
 				mL(2, "metaRolld10", "Not Anchored", "Re-Roll", "Could not find the chat message to update", messageId);
 				return;
 			}
-			//const updatedRoll = await rolld10.toJSON();
 			const renderedRoll = await rolld10.render();
 			//? Call Dice So Nice to show the roll
 			if (game.dice3d && dice > 0) {
@@ -241,153 +224,13 @@ export async function metaRolld10(
 			}
 			chatMessage.update({
 				flavor: enrichedMessage,
-				//rolls: updatedRoll,
 				content: renderedRoll,
 				rollMode: game.settings.get("core", "rollMode"),
 				flags: { metanthropes: { actoruuid: actor.uuid } },
 			});
 		}
+		mL(3, "metaRolld10", "Finished for:", actor.name + "'s", what);
 	}
-	//! removing the else section
-	// else {
-	// 	//* Roll is anchored (shown within an existing message)
-	// 	if (!reroll) {
-	// 		//! Anchored, not re-roll means we'll be creating a new chat message
-	// 		//however we return the roll to anchor here, so when does the message get created?
-	// 		//!atn here (!reroll) returns truthy if reroll==="false" cause it's a string! giati edw kanw parse to string directly oxi opws ta reroll functions
-	// 		mL(3, "metaRolld10", "Anchored", "No Re-roll");
-	// 		//* We store in the dataset all info to display the chat message if needed from rerolls
-	// 		//*? don't print a chat message ?what is reroll exactly? todo: rename to more clean purpose
-	// 		//! do I need this anymore for rerolls to show for all players?
-	// 		// when roll is anchored we don't want DSN to trigger before showing the result.
-	// 		// the resulting dsn animation should show up when the content message does from the parent caller (metaexecute in most cases)
-	// 		//? Call Dice So Nice to show the roll
-	// 		//! enabling causes dice to show in first activation
-	// 		// if (game.dice3d && dice > 0) {
-	// 		// 	await game.dice3d.showForRoll(rolld10, game.user, true, null, false, messageId);
-	// 		// }
-	// 		//! I need to pass the roll object for DSN through the anchor dataset to the damage/healing rerolls so they can trigger the DNS animation when they show the chat message after the initial destiny reroll
-	// 		//const updatedRoll = JSON.stringify(rolld10.toJSON());
-	// 		const updatedRoll = JSON.stringify(foundry.utils.deepClone(rolld10));
-	// 		//const renderedRoll = await rolld10.render();
-	// 		//mL(5, "metaRolld10", "updatedRoll", updatedRoll, "rolld10", rolld10);
-	// 		mL(3, "metaRolld10", "Finished for:", actor.name + "'s", what);
-	// 		return rolld10.toAnchor({
-	// 			label: what,
-	// 			dataset: {
-	// 				total: rollTotal,
-	// 				actoruuid: actor.uuid,
-	// 				item: itemName,
-	// 				what: what,
-	// 				destinyReRoll: destinyReRoll,
-	// 				dice: dice,
-	// 				baseNumber: baseNumber,
-	// 				isHalf: isHalf,
-	// 				reroll: true, //edw
-	// 				rerollCounter: rerollCounter,
-	// 				flavor: enrichedMessage,
-	// 				//content: renderedRoll, //? not including the rendered roll doesn't trigger DSN?
-	// 				rolls: updatedRoll,
-	// 			},
-	// 		});
-	// 	} else {
-	// 		//* Re rolling for an anchor
-	// 		mL(3, "metaRolld10", "Anchored", "Re-Roll");
-	// 		const updatedRoll = JSON.stringify(rolld10.toJSON());
-	// 		const renderedRoll = await rolld10.render();
-
-	// 		const chatData = {
-	// 			speaker: ChatMessage.getSpeaker({ actor: actor }),
-	// 			flavor: enrichedMessage,
-	// 			rolls: updatedRoll,
-	// 			content: renderedRoll,
-	// 			rollMode: game.settings.get("core", "rollMode"),
-	// 			flags: { metanthropes: { actoruuid: actor.uuid } },
-	// 		};
-	// 		const chatMessage = game.messages.get(messageId);
-	// 		if (!chatMessage) {
-	// 			//* If no previous chat message to replace
-	// 			//? This should only happen if it's the first reroll after activation
-	// 			//todo we should not send a new chat message, instead only return anchor?
-	// 			mL(
-	// 				4,
-	// 				"metaRolld10",
-	// 				"Anchored",
-	// 				"Re-Roll",
-	// 				"Could not find the chat message to update",
-	// 				chatMessage,
-	// 				messageId,
-	// 				"Creating new chat message, not returning to anchor",
-	// 			);
-	// 			metanthropes.applications.MetaChatMessage.create(chatData); // ayto kanei to "The Protagonist" message
-	// 			//? AND return the anchor, setting it to false so if we have another reroll we'll update that new message
-	// 			//? xreiazomai na kanw return to anchor giati ayto kanei create to dataset poy thelw gia to rolledDice
-	// 			//? isws auto einai edw poy prepei na zhtaw na min to kanei animate to DSN?
-	// 			mL(3, "metaRolld10", "Finished for:", actor.name + "'s", what);
-	// 			//! issue edw einai oti to Anchor kanei create message enw den tha eprepe.
-	// 			//! omws einai to metaDamageReroll poy kanei create to message me ayto to anchor? giati edw kanoyme return, se poion kanoume return?
-	// 			// todo kai confirm ta dataset params
-	// 			// return rolld10.toAnchor({
-	// 			// 	//ayto kanei to 'The Narrator' message
-	// 			// 	label: what,
-	// 			// 	dataset: {
-	// 			// 		total: rollTotal,
-	// 			// 		actoruuid: actor.uuid,
-	// 			// 		item: itemName,
-	// 			// 		what: what,
-	// 			// 		destinyReRoll: destinyReRoll,
-	// 			// 		dice: dice,
-	// 			// 		baseNumber: baseNumber,
-	// 			// 		isHalf: isHalf,
-	// 			// 		reroll: reroll,
-	// 			// 		anchor: false,
-	// 			// 		rerollCounter: rerollCounter,
-	// 			// 		flavor: enrichedMessage,
-	// 			// 		content: renderedRoll,
-	// 			// 		rolls: updatedRoll,
-	// 			// 	},
-	// 			// });
-	// 		} else {
-	// 			//* Replacing previous chat message
-	// 			mL(3, "metaRolld10", "Anchored", "Re-Roll", "Updating messageId", messageId);
-	// 			//await chatMessage.update(chatData);
-	// 			mL(3, "metaRolld10", "Finished for:", actor.name + "'s", what);
-	// 			//? DSN pleon redundant edw mias kai kanei pickup ta chat message updates?
-	// 			// if (game.dice3d && dice > 0) {
-	// 			// 	await game.dice3d.showForRoll(rolld10, game.user, true, null, false, messageId);
-	// 			// }
-	// 			// //? Call Dice So Nice to show the roll - ayto to xreiazomai gia ta re-rolls
-	// 			// if (game.dice3d && dice > 0) {
-	// 			// 	game.dice3d.showForRoll(rolld10, game.user, true, null, false, messageId);
-	// 			// }
-	// 			// //? Call Dice So Nice to show the roll - ayto to xreiazomai gia ta re-rolls
-	// 			//! removing the bellow solved the 2nd roll bug?
-	// 			// if (game.dice3d && dice > 0) {
-	// 			// 	game.dice3d.showForRoll(rolld10, game.user, true, null, false, messageId);
-	// 			// }
-	// 			return rolld10.toAnchor({
-	// 				label: what,
-	// 				dataset: {
-	// 					total: rollTotal,
-	// 					actoruuid: actor.uuid,
-	// 					item: itemName,
-	// 					what: what,
-	// 					destinyReRoll: destinyReRoll,
-	// 					dice: dice,
-	// 					baseNumber: baseNumber,
-	// 					isHalf: isHalf,
-	// 					reroll: reroll,
-	// 					rerollCounter: rerollCounter,
-	// 					flavor: enrichedMessage,
-	// 					content: renderedRoll,
-	// 					rolls: updatedRoll,
-	// 					//! do I need messageId here?
-	// 				},
-	// 			});
-	// 		}
-	// 	}
-	// }
-	mL(3, "metaRolld10", "Finished for:", actor.name + "'s", what);
 }
 
 /**
@@ -406,6 +249,7 @@ export async function metaRolld10(
 export async function metaRolld10ReRoll(event) {
 	const mL = metanthropes.utils.metaLog;
 	event.preventDefault();
+	//* Gather params
 	const button = event.target;
 	//? Traverse up the DOM to find the parent <li> element with the data-message-id attribute
 	const messageElement = button.closest("li.chat-message");
@@ -445,9 +289,7 @@ export async function metaRolld10ReRoll(event) {
 	}
 	await actor.applyDestinyChange(-1);
 	//* Call metaRolld10
-
-	mL(1, "metaRolld10ReRoll", "Finished, calling metaRolld10", reroll ? "Reroll" : "No reroll");
-
+	mL(3, "metaRolld10ReRoll", "Finished, calling metaRolld10", reroll ? "Reroll" : "No reroll");
 	await metanthropes.dice.metaRolld10(
 		actor,
 		what,
@@ -461,438 +303,4 @@ export async function metaRolld10ReRoll(event) {
 		rerollCounter,
 		messageId,
 	);
-}
-
-/**
- * metaDamageReRoll handles the re-rolling of damage and application of the new damage to targets.
- *
- * @export
- * @async
- * @param {*} event
- * @returns {*}
- */
-export async function metaDamageReRoll(event) {
-	const mL = metanthropes.utils.metaLog;
-	mL(3, "metaDamageReRoll", "Rerolling Damage Result");
-	event.preventDefault();
-	const button = event.target;
-	//? Traverse up the DOM to find the parent <li> element with the data-message-id attribute
-	const messageElement = button.closest("li.chat-message");
-	if (!messageElement) {
-		ui.notifications.warn("Could not find the chat message element.");
-		return;
-	}
-	//? Retrieve the message ID from the data-message-id attribute
-	///pairnw result, alla giati to xreiazomai?
-	const messageId = messageElement.dataset.messageId;
-	if (!messageId) {
-		ui.notifications.warn("Could not retrieve the message ID.");
-		return;
-	}
-	const actoruuid = button.dataset.actoruuid;
-	const what = button.dataset.what;
-	const destinyReRoll = button.dataset.destinyReRoll === "true" ? true : false;
-	const itemName = button.dataset.itemName === "null" ? null : button.dataset.itemName;
-	const isHalf = button.dataset.isHalf === "true" ? true : false;
-	const anchor = button.dataset.anchor === "true" ? true : false;
-	let reroll = button.dataset.reroll === "false" ? false : true;
-	let rerollCounter = parseInt(button.dataset.rerollCounter) || 0;
-	const actor = await fromUuid(actoruuid);
-	const targetedActors = button.dataset.targets ? button.dataset.targets.split(",") : ([] ?? null);
-	//todo ** tha mporousa na to kanw set null kai meta na min kanw call metaRoll ektos ean perimenw result
-	const damageDiceCosmic = parseInt(button.dataset.diceCosmic) || 0;
-	const damageDiceElemental = parseInt(button.dataset.diceElemental) || 0;
-	const damageDiceMaterial = parseInt(button.dataset.diceMaterial) || 0;
-	const damageDicePsychic = parseInt(button.dataset.dicePsychic) || 0;
-	const damageBaseCosmic = parseInt(button.dataset.baseCosmic) || 0;
-	const damageBaseElemental = parseInt(button.dataset.baseElemental) || 0;
-	const damageBaseMaterial = parseInt(button.dataset.baseMaterial) || 0;
-	const damageBasePsychic = parseInt(button.dataset.basePsychic) || 0;
-	const damageSelectedTargets = button.dataset.damageSelectedTargets === "true" ? true : false;
-	const firstMessage = button.dataset?.firstMessage === "true" ? true : false;
-	let contentMessage = ``;
-	let startMessage = ``;
-	let flavorMessage = ``;
-	let rollData = null;
-	//* Return conditions
-	//? Need to check if actor has enough Destiny to spend, because they might have already spent it on another secondary button
-	if (!(actor.currentDestiny > 0 && destinyReRoll)) {
-		ui.notifications.warn(actor.name + " does not have enough Destiny to spend for reroll!");
-		mL(1, "metaDamageReRoll", "Not enough Destiny to spend", "OR", "destinyReRoll is not allowed");
-		button.classList.remove("disabled");
-		return;
-	}
-	//todo giati exw ayto edw, eixa skopo na kanw restructure me vasi ta reroll state?
-	//! reroll = true;
-	//? Need to ensure we have valid targets before reducing destiny
-	if (!targetedActors.length && damageSelectedTargets) {
-		ui.notifications.warn("You must select valid targets first");
-		button.classList.remove("disabled");
-		return;
-	}
-	await actor.applyDestinyChange(-1);
-	//todo see ** above
-	//todo den kanw to roll ean den xreiazetai?
-	let cosmicDamageRollResult = null;
-	let damageCosmicMessage = null;
-	let elementalDamageRollResult = null;
-	let damageElementalMessage = null;
-	let materialDamageRollResult = null;
-	let damageMaterialMessage = null;
-	let psychicDamageRollResult = null;
-	let damagePsychicMessage = null;
-	let rolledDice = [];
-	let dsnRolls = [];
-	//* Cosmic
-	if (damageDiceCosmic > 0 || damageBaseCosmic > 0) {
-		const cosmicDamageRoll = await metanthropes.dice.metaRolld10(
-			actor,
-			`Cosmic Damage`,
-			true,
-			damageDiceCosmic,
-			itemName,
-			damageBaseCosmic,
-			false,
-			true,
-			firstMessage ? false : true,
-			0,
-			firstMessage ? null : messageId,
-		);
-		mL(3, "metaDamageReRoll", "Cosmic Damage Dataset", cosmicDamageRoll.dataset);
-		cosmicDamageRollResult = Number(cosmicDamageRoll.dataset.total);
-		damageCosmicMessage = `${cosmicDamageRoll.outerHTML}`;
-		const cosmicRollData = JSON.parse(cosmicDamageRoll.dataset.rolls);
-		rolledDice.push(cosmicRollData);
-		//? rebuild a real Roll object for DSN from the RollData
-		const cosmicDSNRoll = Roll.fromData(cosmicRollData);
-		dsnRolls.push(cosmicDSNRoll);
-	}
-	//* Elemental
-	if (damageDiceElemental > 0 || damageBaseElemental > 0) {
-		const elementalDamageRoll = await metanthropes.dice.metaRolld10(
-			actor,
-			`Elemental Damage`,
-			true,
-			damageDiceElemental,
-			itemName,
-			damageBaseElemental,
-			false,
-			true,
-			firstMessage ? false : true, //!giati pernaw false edw? //todo kai gia ta alla damage types -- me false paizei swsta to 1o reroll alla den paizoune ta sub-sequent
-			0, //! kai giati to counter einai 0 edw alla seems to be working fine?
-			firstMessage ? null : messageId, //null, //? we don't pass the messageId as that would replace the original message
-		);
-		mL(3, "metaDamageReRoll", "Elemental Damage Dataset", elementalDamageRoll.dataset);
-		elementalDamageRollResult = Number(elementalDamageRoll.dataset.total);
-		damageElementalMessage = `${elementalDamageRoll.outerHTML}`;
-		const elementalRollData = JSON.parse(elementalDamageRoll.dataset.rolls);
-		rolledDice.push(elementalRollData);
-		//? rebuild a real Roll object for DSN from the RollData
-		const elementalDSNRoll = Roll.fromData(elementalRollData);
-		dsnRolls.push(elementalDSNRoll);
-	}
-	//* Material
-	if (damageDiceMaterial > 0 || damageBaseMaterial > 0) {
-		const materialDamageRoll = await metanthropes.dice.metaRolld10(
-			actor,
-			`Material Damage`,
-			true,
-			damageDiceMaterial,
-			itemName,
-			damageBaseMaterial,
-			false,
-			true,
-			firstMessage ? false : true,
-			0,
-			firstMessage ? null : messageId,
-		);
-		mL(3, "metaDamageReRoll", "Material Damage Dataset", materialDamageRoll.dataset);
-		materialDamageRollResult = Number(materialDamageRoll.dataset.total);
-		damageMaterialMessage = `${materialDamageRoll.outerHTML}`;
-		const materialRollData = JSON.parse(materialDamageRoll.dataset.rolls);
-		rolledDice.push(materialRollData);
-		const materialDSNRoll = Roll.fromData(materialRollData);
-		dsnRolls.push(materialDSNRoll);
-	}
-	//* Psychic
-	if (damageDicePsychic > 0 || damageBasePsychic > 0) {
-		const psychicDamageRoll = await metanthropes.dice.metaRolld10(
-			actor,
-			`Psychic Damage`,
-			true,
-			damageDicePsychic,
-			itemName,
-			damageBasePsychic,
-			false,
-			true,
-			firstMessage ? false : true,
-			0,
-			firstMessage ? null : messageId,
-		);
-		mL(3, "metaDamageReRoll", "Psychic Damage Dataset", psychicDamageRoll.dataset);
-		psychicDamageRollResult = Number(psychicDamageRoll.dataset.total);
-		damagePsychicMessage = `${psychicDamageRoll.outerHTML}`;
-		const psychicRollData = JSON.parse(psychicDamageRoll.dataset.rolls);
-		rolledDice.push(psychicRollData);
-		const psychicDSNRoll = Roll.fromData(psychicRollData);
-		dsnRolls.push(psychicDSNRoll);
-	}
-	if (damageSelectedTargets) {
-		//* Work through targets to restore previous Life before applying new Damage
-		const resolvedActors = targetedActors.map((uuid) => fromUuidSync(uuid)).filter(Boolean);
-		await Promise.all(resolvedActors.map((actor) => actor.undoLastLifeChange()));
-		await metanthropes.logic.metaApplyDamage(
-			targetedActors,
-			cosmicDamageRollResult,
-			elementalDamageRollResult,
-			materialDamageRollResult,
-			psychicDamageRollResult,
-		);
-	}
-	//* Build the chat Message
-	if (!reroll) {
-		startMessage = "Rolls again for";
-	} else {
-		startMessage = "Re-Rolls";
-		rerollCounter++;
-		if (rerollCounter > 1) startMessage += ` (@METAFA(xmark, null)${rerollCounter})`;
-	}
-	flavorMessage = `${startMessage} ${itemName}'s @METAFA(burst) Damage`;
-	if (damageSelectedTargets)
-		flavorMessage += ` to ${targetedActors.length} target${targetedActors.length > 1 ? "s" : ""}:<br>`;
-	else flavorMessage += `.<br>`;
-	if (cosmicDamageRollResult > 0)
-		contentMessage += `<div class="meta-roll-inline-results">${damageCosmicMessage}</div>`;
-	if (elementalDamageRollResult > 0)
-		contentMessage += `<div class="meta-roll-inline-results">${damageElementalMessage}</div>`;
-	if (materialDamageRollResult > 0)
-		contentMessage += `<div class="meta-roll-inline-results">${damageMaterialMessage}</div>`;
-	if (psychicDamageRollResult > 0)
-		contentMessage += `<div class="meta-roll-inline-results">${damagePsychicMessage}</div>`;
-	if (actor.currentDestiny > 0) {
-		const damageReRollButton = `<div class="hide-button hidden">
-		<button class="metanthropes-secondary-chat-button damage roll-damage-reroll chat-button-anchor"
-		data-targets="${targetedActors}" data-actoruuid="${actor.uuid}" data-item-name="${itemName}"
-		data-what="Damage" data-anchor="true" data-reroll="true" data-reroll-counter="${rerollCounter}" data-message-id="${messageId}"
-		data-destiny-re-roll="true" data-dice-cosmic="${damageDiceCosmic}" data-base-cosmic="${damageBaseCosmic}"
-		data-dice-elemental="${damageDiceElemental}" data-base-elemental="${damageBaseElemental}"
-		data-dice-material="${damageDiceMaterial}" data-base-material="${damageBaseMaterial}"
-		data-dice-psychic="${damageDicePsychic}" data-base-psychic="${damageBasePsychic}"
-		data-damage-selected-targets="${damageSelectedTargets}"
-		>Spend @METAFA(hand-fingers-crossed) to Reroll @METAFA(burst) Damage
-		</button></div>`;
-		contentMessage += `<hr />`;
-		contentMessage += damageReRollButton;
-		contentMessage += `<br>`;
-	}
-	contentMessage += `<div>${actor.name} has ${actor.currentDestiny} @METAFA(hand-fingers-crossed) Destiny remaining.<br><br></div>`;
-	const enrichedContent = await foundry.applications.ux.TextEditor.enrichHTML(contentMessage, { async: true });
-	const enrichedFlavor = await foundry.applications.ux.TextEditor.enrichHTML(flavorMessage, { async: true });
-	//const updatedRoll = JSON.stringify(.toJSON();
-	//const renderedRoll = await hungerRoll.render();
-	//mL(5, "metaDamageReRoll", "rolls", rolledDice);
-	let chatData = {
-		speaker: ChatMessage.getSpeaker({ actor: actor }),
-		//user: game.user.id,
-		flavor: enrichedFlavor,
-		//rolls: rolledDice, //!enabling causes double rolls on first roll - does not solve the x2+ rerolls from showing though
-		content: enrichedContent,
-		rollMode: game.settings.get("core", "rollMode"),
-		flags: { metanthropes: { actoruuid: actor.uuid } },
-	};
-	if (reroll) {
-		//todo need to figure out how to wait for 3d dice before showing updated message
-		mL(4, "metaDamageReRoll", "Reroll");
-
-		const chatMessage = game.messages.get(messageId);
-		if (!chatMessage) {
-			ui.notifications.warn("Could not find the damage reroll chat message to update.");
-			mL(2, "metaDamageReRoll", "Could not find messageId", messageId);
-			return;
-		}
-
-		//mL(5, "dice", rolledDice);
-		//! removed chatData = { ...chatData, rolls: rolledDice };
-		if (game.dice3d) {
-			for (const roll of dsnRolls) await game.dice3d.showForRoll(roll, game.user, true, null, false, messageId);
-			//await game.dice3d.showForRoll(rolledDice, game.user, true, null, false, messageId);
-		}
-		await chatMessage.update(chatData);
-		//if (game.dice3d) await game.dice3d.waitFor3DAnimationByMessageID(messageId);
-	} else {
-		mL(5, "metaDamageReRoll", "No Reroll");
-		//? disabling DSN animation for this message (it appears from the anchor)
-		//chatData = { ...chatData, flags: { "dice-so-nice": { skip: true } } };
-		//todo fix this causes the double dice appearing on x2 roll?
-		await metanthropes.applications.MetaChatMessage.create(chatData);
-	}
-	mL(
-		3,
-		"metaDamageReRoll",
-		"Applied Damage ReRoll",
-		cosmicDamageRollResult,
-		elementalDamageRollResult,
-		materialDamageRollResult,
-		psychicDamageRollResult,
-		`Target${targetedActors.length > 1 ? "s" : ""}:`,
-		targetedActors.length,
-	);
-}
-
-/**
- * metaHealingReRoll handles the re-rolling of healing and applying the new value to targets.
- *
- * @export
- * @async
- * @param {*} event
- * @returns {*}
- */
-export async function metaHealingReRoll(event) {
-	const mL = metanthropes.utils.metaLog;
-	mL(3, "metaHealingReRoll", "Rerolling Healing Result");
-	event.preventDefault();
-	const button = event.target;
-	//? Traverse up the DOM to find the parent <li> element with the data-message-id attribute
-	const messageElement = button.closest("li.chat-message");
-	if (!messageElement) {
-		ui.notifications.warn("Could not find the chat message element.");
-		return;
-	}
-	//? Retrieve the message ID from the data-message-id attribute
-	const messageId = messageElement.dataset.messageId;
-	if (!messageId) {
-		ui.notifications.warn("Could not retrieve the message ID.");
-		return;
-	}
-	const actoruuid = button.dataset.actoruuid;
-	const what = button.dataset.what;
-	const destinyReRoll = button.dataset.destinyReRoll === "true" ? true : false;
-	const itemName = button.dataset.itemName === "null" ? null : button.dataset.itemName;
-	const isHalf = button.dataset.isHalf === "true" ? true : false;
-	const anchor = button.dataset.anchor === "true" ? true : false;
-	let reroll = button.dataset.reroll === "false" ? false : true;
-	let rerollCounter = parseInt(button.dataset.rerollCounter) ?? 0;
-	const actor = await fromUuid(actoruuid);
-	const targetedActors = button.dataset.targets ? button.dataset.targets.split(",") : ([] ?? null);
-	const healingDice = parseInt(button.dataset.healingDice) ?? 0;
-	const healingBase = parseInt(button.dataset.healingBase) ?? 0;
-	const healSelectedTargets = button.dataset.healSelectedTargets === "true" ? true : false;
-	const firstMessage = button.dataset?.firstMessage === "true" ? true : false;
-	let contentMessage = ``;
-	let startMessage = ``;
-	let flavorMessage = ``;
-	//* Return conditions
-	//? Need to check if actor has enough Destiny to spend, because they might have already spent it on another secondary button
-	if (!(actor.currentDestiny > 0 && destinyReRoll)) {
-		ui.notifications.warn(actor.name + " does not have enough Destiny to spend for reroll!");
-		mL(1, "metaHealingReRoll", "Not enough Destiny to spend", "OR", "destinyReRoll is not allowed");
-		button.classList.remove("disabled");
-		return;
-	}
-	//? Need to ensure we have valid targets before reducing destiny
-	if (!targetedActors.length && healSelectedTargets) {
-		ui.notifications.warn("You must select valid targets first");
-		button.classList.remove("disabled");
-		return;
-	}
-	//! reroll = true;
-	await actor.applyDestinyChange(-1);
-	//* Healing
-	const healingRoll = await metanthropes.dice.metaRolld10(
-		actor,
-		`Healing`,
-		true,
-		healingDice,
-		itemName,
-		healingBase,
-		false,
-		true,
-		firstMessage ? false : true,
-		0,
-		null,
-	);
-	const healingRollResult = Number(healingRoll.dataset.total);
-	const healingMessage = `${healingRoll.outerHTML}<br>`;
-	const healingRollData = JSON.parse(healingRoll.dataset.rolls);
-	const healingDSNRoll = Roll.fromData(healingRollData);
-	if (healSelectedTargets) {
-		//* Work through targets to restore previous Life before applying new Healing
-		const resolvedActors = targetedActors.map((uuid) => fromUuidSync(uuid)).filter(Boolean);
-		await Promise.all(resolvedActors.map((actor) => actor.undoLastLifeChange()));
-		await metanthropes.logic.metaApplyHealing(targetedActors, healingRollResult);
-	}
-	//* Chat Message
-	if (!reroll) {
-		startMessage = "Rolls again for";
-	} else {
-		startMessage = "Re-Rolls";
-		rerollCounter++;
-		if (rerollCounter > 1) startMessage += ` (@METAFA(xmark, null)${rerollCounter})`;
-	}
-	flavorMessage = `${startMessage} ${itemName}'s @METAFA(heart-pulse) Healing`;
-	if (healSelectedTargets)
-		flavorMessage += ` to ${targetedActors.length} target${targetedActors.length > 1 ? "s" : ""}:<br>`;
-	else flavorMessage += `.<br>`;
-	contentMessage += `<div class="meta-roll-inline-results">`;
-	contentMessage += `${healingMessage}`;
-	contentMessage += `</div>`;
-	if (actor.currentDestiny > 0) {
-		const healingReRollButton = `<div class="hide-button hidden">
-		<button class="metanthropes-secondary-chat-button damage roll-healing-reroll chat-button-anchor"
-		data-targets="${targetedActors}" data-actoruuid="${actor.uuid}" data-item-name="${itemName}"
-		data-what="Healing" data-anchor="true" data-reroll="true" data-reroll-counter="${rerollCounter}" data-message-id="${messageId}"
-		data-destiny-re-roll="true" data-healing-dice="${healingDice}" data-healing-base="${healingBase}"
-		data-heal-selected-targets="${healSelectedTargets}"
-		>Spend @METAFA(hand-fingers-crossed) to Reroll @METAFA(heart-pulse) Healing
-		</button></div>`;
-		contentMessage += `<hr />`;
-		contentMessage += healingReRollButton;
-		contentMessage += `<br>`;
-	}
-	contentMessage += `<div>${actor.name} has ${actor.currentDestiny} @METAFA(hand-fingers-crossed) Destiny remaining.<br><br></div>`;
-	const enrichedContent = await foundry.applications.ux.TextEditor.enrichHTML(contentMessage, { async: true });
-	const enrichedFlavor = await foundry.applications.ux.TextEditor.enrichHTML(flavorMessage, { async: true });
-	const finalMessage = `${enrichedFlavor}${enrichedContent}`;
-	//! parsing the object from the string stored in the dataset
-	//todo: see https://gitlab.com/riccisi/foundryvtt-dice-so-nice/-/wikis/Integration#step-3-create-a-chat-message-with-the-required-data
-	const updatedRoll = await JSON.parse(healingRoll.dataset.rolls);
-	const renderedRoll = healingRoll.dataset.content;
-	//mL(2, "healingRoll", updatedRoll, renderedRoll);
-	//const updatedRoll = await healingRoll.toJSON(); //!!!
-	//const renderedRoll = await healingRoll.render();
-	let chatData = {
-		speaker: ChatMessage.getSpeaker({ actor: actor }),
-		flavor: enrichedFlavor,
-		//? deprecated in v12: type: CONST.CHAT_MESSAGE_STYLES.ROLL,
-		//rolls: healingDSNRoll,
-		content: enrichedContent,
-		rollMode: game.settings.get("core", "rollMode"),
-		flags: { metanthropes: { actoruuid: actor.uuid } },
-	};
-	if (reroll) {
-		mL(1, "metaHealingReRoll", "Reroll");
-		//? need to call dsn as we don't use the roll class to update the message
-		// if (game.dice3d) {
-		// 	await game.dice3d.showForRoll(updatedRoll, game.user, true, null, false, messageId);
-		// }
-		const chatMessage = game.messages.get(messageId);
-		if (!chatMessage) {
-			ui.notifications.warn("Could not find the healing reroll chat message to update.");
-			mL(2, "metaHealingReRoll", "Could not find messageId", messageId);
-			return;
-		}
-		if (game.dice3d) {
-			await game.dice3d.showForRoll(healingDSNRoll, game.user, true, null, false, messageId);
-		}
-		//chatMessage.applyRollMode(chatData, "roll");
-		await chatMessage.update(chatData);
-	} else {
-		mL(1, "metaHealingReRoll", "No Reroll");
-		// if (game.dice3d) {
-		// 	await game.dice3d.showForRoll(healingDSNRoll, game.user, true, null, false, messageId);
-		// }
-		//chatData = {...chatData, rolls: healingDSNRoll};
-		await metanthropes.applications.MetaChatMessage.create(chatData);
-	}
-	mL(3, "metaHealingReRoll", "Applied Healing ReRoll", healingRollResult, "# of targets:", targetedActors.length);
 }
